LITE-27471 disable autocomplete selection on keydwon and keyup
