+++ .eslintrc.yaml
@@ -58,6 +58,9 @@ rules:
   require-await:
     - error
 
+ignorePatterns: 
+  - libs
+
 parserOptions:
   parser: "@babel/eslint-parser"
   requireConfigFile: false
