+++ connect_transformations/static/transformations/currency_conversion.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency Conversion</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/currency_conversion.8554bfeaf3bbabd3fb90.js"></script><link href="../transformations/currency_conversion.15f999ee385ccc3cdaea.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><form name="convertCurrency" class="convert-currency"><div class="convert-currency__input-group"><div class="convert-currency__column convert-currency__input"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="convert-currency__input"><label for="from-currency">From Currency</label> <select name="fromCurrency" id="from-currency"></select></div></div><div class="convert-currency__input-group"><div class="convert-currency__column convert-currency__input"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div><div class="convert-currency__input"><label for="to-currency">To Currency</label> <select name="toCurrency" id="to-currency"></select></div></div></form></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency Conversion</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/currency_conversion.b9e3d72f96b5705fafaf.js"></script><link href="../transformations/currency_conversion.15f999ee385ccc3cdaea.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><form name="convertCurrency" class="convert-currency"><div class="convert-currency__input-group"><div class="convert-currency__column convert-currency__input"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="convert-currency__input"><label for="from-currency">From Currency</label> <select name="fromCurrency" id="from-currency"></select></div></div><div class="convert-currency__input-group"><div class="convert-currency__column convert-currency__input"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div><div class="convert-currency__input"><label for="to-currency">To Currency</label> <select name="toCurrency" id="to-currency"></select></div></div></form></div></div></body></html>
