+++ connect_transformations/static/transformations/filter_row.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Filter Row</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/filter_row.c28150b9e26d128d1347.js"></script><link href="../transformations/filter_row.9e2b7d8cadaada9f8c93.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="value">Value</label> <input id="value" class="input" placeholder="Value"/></div></div><div class="button-container"><button id="add" class="button">ADD VALUE</button></div><div class="input-group"><p class="radio-buttons">How to filter rows by condition:</p><div class="radio-buttons"><input type="radio" id="match" name="filter_condition" value="match"/> <label class="label" for="match">By match</label> <input type="radio" id="mismatch" name="filter_condition" value="mismatch"/> <label class="label" for="mismatch">By mismatch</label></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Filter Row</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/filter_row.065285718728bf569d85.js"></script><link href="../transformations/filter_row.9e2b7d8cadaada9f8c93.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="value">Value</label> <input id="value" class="input" placeholder="Value"/></div></div><div class="button-container"><button id="add" class="button">ADD VALUE</button></div><div class="input-group"><p class="radio-buttons">How to filter rows by condition:</p><div class="radio-buttons"><input type="radio" id="match" name="filter_condition" value="match"/> <label class="label" for="match">By match</label> <input type="radio" id="mismatch" name="filter_condition" value="mismatch"/> <label class="label" for="mismatch">By mismatch</label></div></div></div></body></html>
