+++ connect_transformations/static/transformations/lookup_product_item.5c6c84900a4b8cec3e59.js
@@ -2,74 +2,14 @@
 /******/ 	"use strict";
 /******/ 	var __webpack_modules__ = ({
 
-/***/ 616:
+/***/ 179:
 /***/ ((__unused_webpack_module, __unused_webpack___webpack_exports__, __webpack_require__) => {
 
 
-// EXTERNAL MODULE: ../install_temp/node_modules/@cloudblueconnect/connect-ui-toolkit/dist/index.js
-var dist = __webpack_require__(243);
-;// CONCATENATED MODULE: ./ui/src/components.js
-/*
-Copyright (c) 2023, CloudBlue LLC
-All rights reserved.
-*/
+// UNUSED EXPORTS: createOutputColumnForLookup, lookupProductItem
 
-// render UI components - show/hide
-const showComponent = (id) => {
-  if (!id) return;
-  const element = document.getElementById(id);
-  element.classList.remove('hidden');
-};
-
-const hideComponent = (id) => {
-  if (!id) return;
-  const element = document.getElementById(id);
-  element.classList.add('hidden');
-};
-
-const showError = (message) => {
-  const oldError = document.getElementById('error');
-  if (oldError) {
-    oldError.remove();
-  }
-  const error = document.createElement('div');
-  error.id = 'error';
-  error.innerHTML = `<div class="c-alert">${message}</div>`;
-  document.getElementsByTagName('body')[0].appendChild(error);
-  document.getElementById('error').scrollIntoView();
-};
-
-const hideError = () => {
-  const error = document.getElementById('error');
-  if (error) {
-    error.remove();
-  }
-};
-
-const getAddSvg = () => '<svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24"><path d="M0 0h24v24H0z" fill="none"/><path d="M19 13H13v6h-2v-6H5v-2h6V5h2v6h6v2z"/></svg>';
-
-const getDeleteSvg = () => '<svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24"><path d="M0 0h24v24H0z" fill="none"/><path d="M0 0h24v24H0V0z" fill="none"/><path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zm2.46-7.12l1.41-1.41L12 12.59l2.12-2.12 1.41 1.41L13.41 14l2.12 2.12-1.41 1.41L12 15.41l-2.12 2.12-1.41-1.41L10.59 14l-2.13-2.12zM15.5 4l-1-1h-5l-1 1H5v2h14V4z"/></svg>';
-
-const getAddButton = (index) => {
-  const button = document.createElement('button');
-  button.classList.add('add-button');
-  button.id = 'add-button';
-  button.setAttribute('data-row-index', index);
-  button.innerHTML = getAddSvg();
-
-  return button;
-};
-
-const getDeleteButton = (index) => {
-  const button = document.createElement('button');
-  button.classList.add('delete-button');
-  button.id = `delete-${index}`;
-  button.setAttribute('data-row-index', index);
-  button.innerHTML = getDeleteSvg();
-
-  return button;
-};
-
+// EXTERNAL MODULE: ./node_modules/@cloudblueconnect/connect-ui-toolkit/dist/index.js
+var dist = __webpack_require__(164);
 ;// CONCATENATED MODULE: ./ui/src/utils.js
 
 /*
@@ -156,7 +96,69 @@ const getAirtableTables = (key, baseId) => fetch(`/api/airtable_lookup/tables?ap
 }).then((response) => response.json());
 
 
-;// CONCATENATED MODULE: ./ui/src/pages/transformations/airtable_lookup.js
+;// CONCATENATED MODULE: ./ui/src/components.js
+/*
+Copyright (c) 2023, CloudBlue LLC
+All rights reserved.
+*/
+
+// render UI components - show/hide
+const showComponent = (id) => {
+  if (!id) return;
+  const element = document.getElementById(id);
+  element.classList.remove('hidden');
+};
+
+const hideComponent = (id) => {
+  if (!id) return;
+  const element = document.getElementById(id);
+  element.classList.add('hidden');
+};
+
+const showError = (message) => {
+  const oldError = document.getElementById('error');
+  if (oldError) {
+    oldError.remove();
+  }
+  const error = document.createElement('div');
+  error.id = 'error';
+  error.innerHTML = `<div class="c-alert">${message}</div>`;
+  document.getElementsByTagName('body')[0].appendChild(error);
+  document.getElementById('error').scrollIntoView();
+};
+
+const hideError = () => {
+  const error = document.getElementById('error');
+  if (error) {
+    error.remove();
+  }
+};
+
+const getAddSvg = () => '<svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24"><path d="M0 0h24v24H0z" fill="none"/><path d="M19 13H13v6h-2v-6H5v-2h6V5h2v6h6v2z"/></svg>';
+
+const getDeleteSvg = () => '<svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24"><path d="M0 0h24v24H0z" fill="none"/><path d="M0 0h24v24H0V0z" fill="none"/><path d="M6 19c0 1.1.9 2 2 2h8c1.1 0 2-.9 2-2V7H6v12zm2.46-7.12l1.41-1.41L12 12.59l2.12-2.12 1.41 1.41L13.41 14l2.12 2.12-1.41 1.41L12 15.41l-2.12 2.12-1.41-1.41L10.59 14l-2.13-2.12zM15.5 4l-1-1h-5l-1 1H5v2h14V4z"/></svg>';
+
+const getAddButton = (index) => {
+  const button = document.createElement('button');
+  button.classList.add('add-button');
+  button.id = 'add-button';
+  button.setAttribute('data-row-index', index);
+  button.innerHTML = getAddSvg();
+
+  return button;
+};
+
+const getDeleteButton = (index) => {
+  const button = document.createElement('button');
+  button.classList.add('delete-button');
+  button.id = `delete-${index}`;
+  button.setAttribute('data-row-index', index);
+  button.innerHTML = getDeleteSvg();
+
+  return button;
+};
+
+;// CONCATENATED MODULE: ./ui/src/pages/transformations/lookup_product_item.js
 /*
 Copyright (c) 2023, CloudBlue LLC
 All rights reserved.
@@ -170,187 +172,143 @@ All rights reserved.
 
 
 
-const createCopyRow = (parent, index, options, input, output) => {
-  const item = document.createElement('div');
-  item.classList.add('list-wrapper');
-  item.id = `wrapper-${index}`;
-  item.style.width = '100%';
-  item.innerHTML = `
-      <select class="list" style="width: 35%;" ${input ? `value="${input.id}"` : ''}>
-        ${options.map((column) => `
-          <option value="${column.id}" ${input && input.id === column.id ? 'selected' : ''}>
-            ${column.name}
-          </option>`).join(' ')}
-      </select>
-      <input type="text" placeholder="Copy column name" style="width: 35%;" ${output ? `value="${output.name}"` : ''} />
-      <button id="delete-${index}" class="button delete-button">DELETE</button>
-    `;
-  parent.appendChild(item);
+const createOutputColumnForLookup = (prefix, name) => ({
+  name: `${prefix}.${name}`,
+  type: 'string',
+  description: '',
+});
 
-  document.getElementById(`delete-${index}`).addEventListener('click', () => {
-    if (document.getElementsByClassName('list-wrapper').length === 1) {
-      showError('You need to have at least one row');
-    } else {
-      document.getElementById(`wrapper-${index}`).remove();
-      const buttons = document.getElementsByClassName('delete-button');
-      if (buttons.length === 1) {
-        buttons[0].disabled = true;
-      }
-    }
-  });
-  const buttons = document.getElementsByClassName('delete-button');
-  for (let i = 0; i < buttons.length; i += 1) {
-    if (buttons.length === 1) {
-      buttons[i].disabled = true;
-    } else {
-      buttons[i].disabled = false;
-    }
-  }
-};
-
-const createOptions = (selectId, options) => {
-  const select = document.getElementById(selectId);
-  select.innerHTML = `
-        <option disabled selected value>Please select an option</option>
-        ${options.map((column) => `
-          <option value="${column.id}">
-            ${column.name}
-          </option>`).join(' ')}
-    `;
-};
-
-const removeDisabled = selector => document.getElementById(selector).removeAttribute('disabled');
-
-const airtable = (app) => {
+const lookupProductItem = (app) => {
   if (!app) return;
-  hideComponent('loader');
-  showComponent('app');
 
   let columns = [];
-  let airtableColumns = [];
-  let apiKey;
-  let baseId;
-  let tableId;
-  let tables;
-  let mapInputColumn;
-  let mapAirtableColumn;
-  const inputColumnSelect = document.getElementById('input-column-select');
-  const airtableFieldSelect = document.getElementById('field-select');
-  const addButton = document.getElementById('add');
+  const toggleProductId = (value) => {
+    if (value === 'product_id') {
+      hideComponent('product_column_input');
+      showComponent('product_id_input');
+    } else {
+      hideComponent('product_id_input');
+      showComponent('product_column_input');
+    }
+  };
 
-  app.listen('config', (config) => {
-    const baseSelect = document.getElementById('base-select');
-    const content = document.getElementById('content');
-    const tableSelect = document.getElementById('table-select');
-    const keyInput = document.getElementById('key-input');
-    let airtableBases;
-    let rowIndex = 0;
-    columns = config.context.available_columns;
+  app.listen('config', async (config) => {
+    const {
+      context: { available_columns: availableColumns, stream },
+      settings,
+    } = config;
 
-    keyInput.addEventListener('input', async () => {
-      apiKey = keyInput.value;
-      if (apiKey.length < 50) return;
+    const hasProduct = 'product' in stream.context;
+    columns = availableColumns;
+    const criteria = await getLookupProductItemCriteria();
 
-      try {
-        airtableBases = await getAirtableBases(apiKey);
-        if (airtableBases.error) {
-          throw new Error(airtableBases.error);
-        }
-        hideError();
-      } catch (e) {
-        showError(e);
+    // defaults
+    document.getElementById('leave_empty').checked = true;
+    document.getElementById('by_product_id').checked = true;
+    hideComponent('product_column_input');
+    showComponent('product_id_input');
+    hideComponent('loader');
+    showComponent('app');
+
+    Object.keys(criteria).forEach((key) => {
+      const option = document.createElement('option');
+      option.value = key;
+      option.text = criteria[key];
+      document.getElementById('criteria').appendChild(option);
+    });
+
+    availableColumns.forEach((column) => {
+      const option = document.createElement('option');
+      option.value = column.id;
+      option.text = column.name;
+      document.getElementById('column').appendChild(option);
+
+      const anotherOption = document.createElement('option');
+      anotherOption.value = column.id;
+      anotherOption.text = column.name;
+      document.getElementById('product_id_column').appendChild(anotherOption);
+    });
+
+    if (hasProduct === true) {
+      document.getElementById('product_id').value = stream.context.product.id;
+      hideComponent('product_id_input');
+      hideComponent('product_column_input');
+      hideComponent('product_id_radio_group');
+      hideComponent('no_product');
+    }
+
+    if (settings) {
+      document.getElementById('product_id').value = settings.product_id;
+      document.getElementById('criteria').value = settings.lookup_type;
+      document.getElementById('column').value = columns.find((c) => c.name === settings.from).id;
+      document.getElementById('product_id_column').value = columns.find((c) => c.name === settings.product_column).id;
+      document.getElementById('prefix').value = settings.prefix;
+      if (settings.action_if_not_found === 'leave_empty') {
+        document.getElementById('leave_empty').checked = true;
+      } else {
+        document.getElementById('fail').checked = true;
       }
+      if (settings.product_lookup_mode === 'id') {
+        document.getElementById('by_product_id').checked = true;
+        hideComponent('product_column_input');
+        showComponent('product_id_input');
+      } else {
+        document.getElementById('by_product_column').checked = true;
+        hideComponent('product_id_input');
+        showComponent('product_column_input');
+      }
+    }
 
-      createOptions('base-select', airtableBases);
-      removeDisabled('base-select');
-    });
-
-    baseSelect.addEventListener('change', async () => {
-      baseId = baseSelect.value;
-      tables = await getAirtableTables(apiKey, baseId);
-      hideError();
-
-      createOptions('table-select', tables);
-      removeDisabled('table-select');
-    });
-
-    tableSelect.addEventListener('change', () => {
-      tableId = tableSelect.value;
-      const currentTable = tables.find(x => x.id === tableId);
-      airtableColumns = currentTable.columns;
-      hideError();
-
-      createOptions('field-select', airtableColumns);
-      createOptions('input-column-select', columns);
-      removeDisabled('field-select');
-      removeDisabled('input-column-select');
-    });
-
-    inputColumnSelect.addEventListener('change', () => {
-      mapInputColumn = columns.find((column) => column.id === inputColumnSelect.value);
-      if (mapAirtableColumn) removeDisabled('add');
-      hideError();
-    });
-
-    airtableFieldSelect.addEventListener('change', () => {
-      mapAirtableColumn = airtableColumns.find((column) => column.id === airtableFieldSelect.value);
-      if (mapInputColumn) removeDisabled('add');
-      hideError();
-    });
-
-    addButton.addEventListener('click', () => {
-      rowIndex += 1;
-      createCopyRow(content, rowIndex, airtableColumns);
-    });
+    const radios = document.getElementsByName('product_id_radio');
+    for (let i = 0, max = radios.length; i < max; i += 1) {
+      radios[i].onclick = () => {
+        toggleProductId(radios[i].value);
+      };
+    }
   });
 
   app.listen('save', async () => {
-    let overview = '';
-    if (!mapInputColumn || !mapAirtableColumn) {
-      showError('Please complete all the fields');
+    const criteria = document.getElementById('criteria').value;
+    const columnId = document.getElementById('column').value;
+    const prefix = document.getElementById('prefix').value;
+    const column = columns.find((c) => c.id === columnId);
+    const actionIfNotFound = document.getElementById('leave_empty').checked ? 'leave_empty' : 'fail';
+    const productLookupMode = document.getElementById('by_product_id').checked ? 'id' : 'column';
+    const productId = document.getElementById('product_id').value;
+    const productColumnId = document.getElementById('product_id_column').value;
+    const productColumn = columns.find((c) => c.id === productColumnId);
 
-      return;
+    const input = [column];
+    if (productLookupMode === 'column') {
+      input.push(productColumn);
     }
 
     const data = {
       settings: {
-        api_key: apiKey,
-        base_id: baseId,
-        table_id: tableId,
-        map_by: {
-          input_column: mapInputColumn.name,
-          airtable_column: mapAirtableColumn.name,
-        },
-        mapping: [],
+        product_id: productId,
+        lookup_type: criteria,
+        from: column.name,
+        prefix,
+        action_if_not_found: actionIfNotFound,
+        product_column: productColumn?.name ?? '',
+        product_lookup_mode: productLookupMode,
       },
       columns: {
-        input: [mapInputColumn],
-        output: [],
+        input,
+        output: [
+          'product.id',
+          'product.name',
+          'item.id',
+          'item.name',
+          'item.unit',
+          'item.period',
+          'item.mpn',
+        ].map((name) => createOutputColumnForLookup(prefix, name)),
       },
     };
 
-    const form = document.getElementsByClassName('list-wrapper');
-    // eslint-disable-next-line no-restricted-syntax
-    for (const line of form) {
-      const inputId = line.getElementsByTagName('select')[0].value;
-      const outputName = line.getElementsByTagName('input')[0].value;
-
-      const inputColumn = airtableColumns.find((column) => column.id === inputId);
-
-      const outputColumn = {
-        name: outputName,
-        description: '',
-      };
-      const setting = {
-        from: inputColumn.name,
-        to: outputName,
-      };
-      data.settings.mapping.push(setting);
-      data.columns.output.push(outputColumn);
-    }
-
     try {
-      overview = await validate('airtable_lookup', data);
+      const overview = await validate('lookup_product_item', data);
       if (overview.error) {
         throw new Error(overview.error);
       }
@@ -362,8 +320,7 @@ const airtable = (app) => {
 };
 
 (0,dist/* default */.ZP)({ })
-  .then(airtable);
-
+  .then(lookupProductItem);
 
 
 /***/ })
@@ -455,7 +412,7 @@ const airtable = (app) => {
 /******/ 		// undefined = chunk not loaded, null = chunk preloaded/prefetched
 /******/ 		// [resolve, reject, Promise] = chunk loading, 0 = chunk loaded
 /******/ 		var installedChunks = {
-/******/ 			18: 0
+/******/ 			784: 0
 /******/ 		};
 /******/ 		
 /******/ 		// no chunk on demand loading
@@ -505,7 +462,7 @@ const airtable = (app) => {
 /******/ 	// startup
 /******/ 	// Load entry module and return exports
 /******/ 	// This entry module depends on other loaded chunks and execution need to be delayed
-/******/ 	var __webpack_exports__ = __webpack_require__.O(undefined, [216], () => (__webpack_require__(616)))
+/******/ 	var __webpack_exports__ = __webpack_require__.O(undefined, [216], () => (__webpack_require__(179)))
 /******/ 	__webpack_exports__ = __webpack_require__.O(__webpack_exports__);
 /******/ 	
 /******/ })()
