+++ connect_transformations/static/transformations/lookup_product_item.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup Product Item</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/lookup_product_item.99b1e1d82cd7b53abe8a.js"></script><link href="../transformations/lookup_product_item.61e184b682acdd7d3871.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria includes CloudBlue Item MPN and CloudBlue Item ID.</div><div class="input-group"><label class="label" for="criteria">Criteria</label> <select class="input" id="criteria"></select></div><div class="subtitle no-product" id="no_product">Warning: No product associated with this data stream. Please either enter a Product ID, which will be applied to all product item lookups, or select a Product column containing the row-specific product IDs.</div><div class="input-group" id="product_id_radio_group" style="margin-bottom: 8px"><p class="radio-buttons">Please specify a way to find product:</p><div class="radio-buttons"><input type="radio" id="by_product_id" name="product_id_radio" value="product_id"/> <label class="label" for="by_product_id">By static product ID</label><br/><input type="radio" id="by_product_column" name="product_id_radio" value="product_column"/> <label class="label" for="by_product_column">By column containing product IDs</label></div></div><div class="input-group" id="product_id_input"><label class="label" for="product_id">Product ID</label> <input class="input" id="product_id"/></div><div class="input-group" id="product_column_input"><label class="label" for="product_id_column">Product ID Column</label> <select class="input" id="product_id_column"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">What To Do If Product Item Not Found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup Product Item</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_product_item.5c6c84900a4b8cec3e59.js"></script><link href="../transformations/lookup_product_item.61e184b682acdd7d3871.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria includes CloudBlue Item MPN and CloudBlue Item ID.</div><div class="input-group"><label class="label" for="criteria">Criteria</label> <select class="input" id="criteria"></select></div><div class="subtitle no-product" id="no_product">Warning: No product associated with this data stream. Please either enter a Product ID, which will be applied to all product item lookups, or select a Product column containing the row-specific product IDs.</div><div class="input-group" id="product_id_radio_group" style="margin-bottom: 8px"><p class="radio-buttons">Please specify a way to find product:</p><div class="radio-buttons"><input type="radio" id="by_product_id" name="product_id_radio" value="product_id"/> <label class="label" for="by_product_id">By static product ID</label><br/><input type="radio" id="by_product_column" name="product_id_radio" value="product_column"/> <label class="label" for="by_product_column">By column containing product IDs</label></div></div><div class="input-group" id="product_id_input"><label class="label" for="product_id">Product ID</label> <input class="input" id="product_id"/></div><div class="input-group" id="product_column_input"><label class="label" for="product_id_column">Product ID Column</label> <select class="input" id="product_id_column"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">What To Do If Product Item Not Found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
