+++ connect_transformations/static/transformations/lookup_subscription.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup Subscription</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/lookup_subscription.93c6eab83331d3944391.js"></script><link href="../transformations/lookup_subscription.61e184b682acdd7d3871.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria includes Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criteria</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">What To Do If Subscription Not Found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup Subscription</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_subscription.76988913babb7b470db0.js"></script><link href="../transformations/lookup_subscription.61e184b682acdd7d3871.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria includes Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criteria</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">What To Do If Subscription Not Found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
