+++ connect_transformations/static/transformations/manual.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Manual</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/manual.fa642e806d249135316b.js"></script><link href="../transformations/manual.57f7accd9c102ae43e15.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div id="description"><h5>Description:</h5><textarea id="description-text" placeholder="Description text" rows="6"></textarea></div><div id="edit-input"><h5>Input columns:</h5><table id="edit-input-columns-table"><thead><tr><th>ID</th><th>NAME</th><th>TYPE</th><th>DESCRIPTION</th><th></th></tr></thead><tbody id="edit-input-columns"></tbody></table></div><div id="settings"><h5>Settings:</h5><textarea id="settings-text" rows="6">{}</textarea></div><div id="output"><h5>Output columns:</h5><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Manual</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/manual.332571ab83705fe8cbbc.js"></script><link href="../transformations/manual.57f7accd9c102ae43e15.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div id="description"><h5>Description:</h5><textarea id="description-text" placeholder="Description text" rows="6"></textarea></div><div id="edit-input"><h5>Input columns:</h5><table id="edit-input-columns-table"><thead><tr><th>ID</th><th>NAME</th><th>TYPE</th><th>DESCRIPTION</th><th></th></tr></thead><tbody id="edit-input-columns"></tbody></table></div><div id="settings"><h5>Settings:</h5><textarea id="settings-text" rows="6">{}</textarea></div><div id="output"><h5>Output columns:</h5><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></div></div></body></html>
