+++ connect_transformations/static/transformations/split_column.98f063fbd31f1e7fdb89.js
@@ -2,14 +2,14 @@
 /******/ 	"use strict";
 /******/ 	var __webpack_modules__ = ({
 
-/***/ 179:
+/***/ 118:
 /***/ ((__unused_webpack_module, __unused_webpack___webpack_exports__, __webpack_require__) => {
 
 
-// UNUSED EXPORTS: createOutputColumnForLookup, lookupProductItem
+// UNUSED EXPORTS: createGroupRows, splitColumn
 
-// EXTERNAL MODULE: ../install_temp/node_modules/@cloudblueconnect/connect-ui-toolkit/dist/index.js
-var dist = __webpack_require__(243);
+// EXTERNAL MODULE: ./node_modules/@cloudblueconnect/connect-ui-toolkit/dist/index.js
+var dist = __webpack_require__(164);
 ;// CONCATENATED MODULE: ./ui/src/utils.js
 
 /*
@@ -158,7 +158,7 @@ const getDeleteButton = (index) => {
   return button;
 };
 
-;// CONCATENATED MODULE: ./ui/src/pages/transformations/lookup_product_item.js
+;// CONCATENATED MODULE: ./ui/src/pages/transformations/split_column.js
 /*
 Copyright (c) 2023, CloudBlue LLC
 All rights reserved.
@@ -172,145 +172,214 @@ All rights reserved.
 
 
 
-const createOutputColumnForLookup = (prefix, name) => ({
-  name: `${prefix}.${name}`,
-  type: 'string',
-  description: '',
-});
+function getCurrentGroups(parent) {
+  const descendents = parent.getElementsByTagName('input');
+  const currentGroups = {};
+  for (let i = 0; i < descendents.length; i += 1) {
+    const element = descendents[i];
+    const dataType = document.getElementById(`datatype-${element.id}`);
+    if (dataType && dataType.value === 'decimal') {
+      const precision = document.getElementById(`precision-${element.id}`).value;
+      currentGroups[element.id] = {
+        name: element.value,
+        type: dataType.value,
+        precision,
+      };
+    } else {
+      currentGroups[element.id] = { name: element.value, type: dataType.value };
+    }
+  }
 
-const lookupProductItem = (app) => {
+  return currentGroups;
+}
+
+function buildSelectColumnType(groupKey) {
+  return `
+  <select id="datatype-${groupKey}">
+  <option value="string" selected>String</option>
+  <option value="integer">Integer</option>
+  <option value="decimal">Decimal</option>
+  <option value="boolean">Boolean</option>
+  <option value="datetime">Datetime</option>
+  </select>
+  `;
+}
+
+function buildSelectColumnPrecision(groupKey) {
+  return `
+  <select id="precision-${groupKey}">
+  <option value="2" selected>2 decimals</option>
+  <option value="3">3 decimals</option>
+  <option value="4">4 decimals</option>
+  <option value="5">5 decimals</option>
+  <option value="6">6 decimals</option>
+  <option value="7">7 decimals</option>
+  <option value="8">8 decimals</option>
+  </select>
+  `;
+}
+
+function buildGroups(groups) {
+  const parent = document.getElementById('output');
+  parent.innerHTML = '';
+  if (Object.keys(groups).length > 0) {
+    const item = document.createElement('div');
+    item.setAttribute('class', 'wrapper output-header');
+    item.innerHTML = `
+    <div>
+    Name
+    </div>
+    <div>
+    Type
+    </div>
+    <div>
+    Precision
+    </div>
+    `;
+    parent.appendChild(item);
+  }
+  Object.keys(groups).forEach(groupKey => {
+    const groupValue = groups[groupKey];
+    const item = document.createElement('div');
+    item.className = 'wrapper';
+    const selectType = buildSelectColumnType(groupKey);
+    const selectPrecision = buildSelectColumnPrecision(groupKey);
+    item.innerHTML = `
+    <div>
+    <input 
+    type="text" 
+    class="output-input" 
+    id="${groupKey}"
+    placeholder="${groupKey} group name" 
+    value="${groupValue.name}"/>
+    </div>
+    <div>
+    ${selectType}
+    </div>
+    <div>
+    ${selectPrecision}
+    </div>
+    `;
+    parent.appendChild(item);
+    document.getElementById(`datatype-${groupKey}`).value = groupValue.type;
+    const precisionSelect = document.getElementById(`precision-${groupKey}`);
+    precisionSelect.disabled = groupValue.type !== 'decimal';
+    precisionSelect.value = groupValue.precision || '2';
+    document.getElementById(`datatype-${groupKey}`).addEventListener('change', () => {
+      if (document.getElementById(`datatype-${groupKey}`).value === 'decimal') {
+        document.getElementById(`precision-${groupKey}`).disabled = false;
+      } else {
+        document.getElementById(`precision-${groupKey}`).disabled = true;
+      }
+    });
+  });
+}
+
+const createGroupRows = async () => {
+  const parent = document.getElementById('output');
+  const groups = getCurrentGroups(parent);
+  const pattern = document.getElementById('pattern').value;
+  if (pattern) {
+    const body = { pattern, groups };
+    const response = await getGroups(body);
+    if (response.error) {
+      showError(response.error);
+    } else {
+      buildGroups(response.groups);
+    }
+  } else {
+    showError('The regular expression is empty');
+  }
+};
+
+const splitColumn = (app) => {
   if (!app) return;
 
   let columns = [];
-  const toggleProductId = (value) => {
-    if (value === 'product_id') {
-      hideComponent('product_column_input');
-      showComponent('product_id_input');
-    } else {
-      hideComponent('product_id_input');
-      showComponent('product_column_input');
-    }
-  };
 
-  app.listen('config', async (config) => {
+  app.listen('config', (config) => {
     const {
-      context: { available_columns: availableColumns, stream },
+      context: { available_columns: availableColumns },
       settings,
     } = config;
 
-    const hasProduct = 'product' in stream.context;
+    showComponent('loader');
+    hideComponent('app');
+
     columns = availableColumns;
-    const criteria = await getLookupProductItemCriteria();
-
-    // defaults
-    document.getElementById('leave_empty').checked = true;
-    document.getElementById('by_product_id').checked = true;
-    hideComponent('product_column_input');
-    showComponent('product_id_input');
-    hideComponent('loader');
-    showComponent('app');
-
-    Object.keys(criteria).forEach((key) => {
-      const option = document.createElement('option');
-      option.value = key;
-      option.text = criteria[key];
-      document.getElementById('criteria').appendChild(option);
-    });
 
     availableColumns.forEach((column) => {
       const option = document.createElement('option');
       option.value = column.id;
       option.text = column.name;
       document.getElementById('column').appendChild(option);
-
-      const anotherOption = document.createElement('option');
-      anotherOption.value = column.id;
-      anotherOption.text = column.name;
-      document.getElementById('product_id_column').appendChild(anotherOption);
     });
 
-    if (hasProduct === true) {
-      document.getElementById('product_id').value = stream.context.product.id;
-      hideComponent('product_id_input');
-      hideComponent('product_column_input');
-      hideComponent('product_id_radio_group');
-      hideComponent('no_product');
-    }
-
     if (settings) {
-      document.getElementById('product_id').value = settings.product_id;
-      document.getElementById('criteria').value = settings.lookup_type;
-      document.getElementById('column').value = columns.find((c) => c.name === settings.from).id;
-      document.getElementById('product_id_column').value = columns.find((c) => c.name === settings.product_column).id;
-      document.getElementById('prefix').value = settings.prefix;
-      if (settings.action_if_not_found === 'leave_empty') {
-        document.getElementById('leave_empty').checked = true;
-      } else {
-        document.getElementById('fail').checked = true;
-      }
-      if (settings.product_lookup_mode === 'id') {
-        document.getElementById('by_product_id').checked = true;
-        hideComponent('product_column_input');
-        showComponent('product_id_input');
-      } else {
-        document.getElementById('by_product_column').checked = true;
-        hideComponent('product_id_input');
-        showComponent('product_column_input');
-      }
+      document.getElementById('pattern').value = settings.regex.pattern;
+      const columnId = columns.find((c) => c.name === settings.from).id;
+      document.getElementById('column').value = columnId;
+      buildGroups(settings.regex.groups);
     }
 
-    const radios = document.getElementsByName('product_id_radio');
-    for (let i = 0, max = radios.length; i < max; i += 1) {
-      radios[i].onclick = () => {
-        toggleProductId(radios[i].value);
-      };
-    }
+    document.getElementById('refresh').addEventListener('click', () => {
+      createGroupRows();
+    });
+    hideComponent('loader');
+    showComponent('app');
   });
 
   app.listen('save', async () => {
-    const criteria = document.getElementById('criteria').value;
-    const columnId = document.getElementById('column').value;
-    const prefix = document.getElementById('prefix').value;
-    const column = columns.find((c) => c.id === columnId);
-    const actionIfNotFound = document.getElementById('leave_empty').checked ? 'leave_empty' : 'fail';
-    const productLookupMode = document.getElementById('by_product_id').checked ? 'id' : 'column';
-    const productId = document.getElementById('product_id').value;
-    const productColumnId = document.getElementById('product_id_column').value;
-    const productColumn = columns.find((c) => c.id === productColumnId);
+    const data = {
+      settings: {},
+      columns: {
+        input: [],
+        output: [],
+      },
+      overview: '',
+    };
+    showComponent('loader');
+    hideComponent('app');
 
-    const input = [column];
-    if (productLookupMode === 'column') {
-      input.push(productColumn);
+    const inputSelector = document.getElementById('column');
+    const selectedColumn = inputSelector.options[inputSelector.selectedIndex].text;
+    const inputColumn = columns.find((column) => column.id === inputSelector.value);
+    data.columns.input.push(inputColumn);
+
+    const selector = document.getElementById('output');
+    const options = selector.getElementsByTagName('input');
+    for (let i = 0; i < options.length; i += 1) {
+      const option = options[i];
+      const dataType = document.getElementById(`datatype-${option.id}`).value;
+      const outputColumn = {
+        name: option.value,
+        type: dataType,
+        description: '',
+        constraints: {},
+      };
+      if (dataType === 'decimal') {
+        const precision = document.getElementById(`precision-${option.id}`).value;
+        outputColumn.constraints = { precision: parseInt(precision, 10) };
+      }
+      data.columns.output.push(outputColumn);
     }
 
-    const data = {
-      settings: {
-        product_id: productId,
-        lookup_type: criteria,
-        from: column.name,
-        prefix,
-        action_if_not_found: actionIfNotFound,
-        product_column: productColumn?.name ?? '',
-        product_lookup_mode: productLookupMode,
-      },
-      columns: {
-        input,
-        output: [
-          'product.id',
-          'product.name',
-          'item.id',
-          'item.name',
-          'item.unit',
-          'item.period',
-          'item.mpn',
-        ].map((name) => createOutputColumnForLookup(prefix, name)),
+    data.settings = {
+      from: selectedColumn,
+      regex: {
+        pattern: document.getElementById('pattern').value,
+        groups: getCurrentGroups(document.getElementById('output')),
       },
     };
 
     try {
-      const overview = await validate('lookup_product_item', data);
+      const overview = await validate('split_column', data);
       if (overview.error) {
         throw new Error(overview.error);
+      }
+
+      if (data.columns.output.length === 0) {
+        throw new Error('No output columns defined');
       }
       app.emit('save', { data: { ...data, ...overview }, status: 'ok' });
     } catch (e) {
@@ -320,7 +389,7 @@ const lookupProductItem = (app) => {
 };
 
 (0,dist/* default */.ZP)({ })
-  .then(lookupProductItem);
+  .then(splitColumn);
 
 
 /***/ })
@@ -412,7 +481,7 @@ const lookupProductItem = (app) => {
 /******/ 		// undefined = chunk not loaded, null = chunk preloaded/prefetched
 /******/ 		// [resolve, reject, Promise] = chunk loading, 0 = chunk loaded
 /******/ 		var installedChunks = {
-/******/ 			784: 0
+/******/ 			12: 0
 /******/ 		};
 /******/ 		
 /******/ 		// no chunk on demand loading
@@ -462,7 +531,7 @@ const lookupProductItem = (app) => {
 /******/ 	// startup
 /******/ 	// Load entry module and return exports
 /******/ 	// This entry module depends on other loaded chunks and execution need to be delayed
-/******/ 	var __webpack_exports__ = __webpack_require__.O(undefined, [216], () => (__webpack_require__(179)))
+/******/ 	var __webpack_exports__ = __webpack_require__.O(undefined, [216], () => (__webpack_require__(118)))
 /******/ 	__webpack_exports__ = __webpack_require__.O(__webpack_exports__);
 /******/ 	
 /******/ })()
