+++ connect_transformations/static/transformations/split_column.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Split Column</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/split_column.9d12961b3edbf58a5add.js"></script><link href="../transformations/split_column.718e9da13fe7e606d87d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div id="input-column" class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div id="regex-column" class="input-group"><label class="label" for="pattern">Regular Expression</label> <input id="pattern" placeholder="Regular expression"/></div><div id="output-column" class="input-group"><label class="label" for="output">Output Columns</label><div id="output"></div><button type="button" id="refresh" class="button">Validate</button></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Split Column</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/split_column.98f063fbd31f1e7fdb89.js"></script><link href="../transformations/split_column.718e9da13fe7e606d87d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div id="input-column" class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div id="regex-column" class="input-group"><label class="label" for="pattern">Regular Expression</label> <input id="pattern" placeholder="Regular expression"/></div><div id="output-column" class="input-group"><label class="label" for="output">Output Columns</label><div id="output"></div><button type="button" id="refresh" class="button">Validate</button></div></div></div></body></html>
