+++ connect_transformations/static/transformations/vat_rate.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/VAT Rate</title><script defer="defer" src="../vendors.65cbf3fa909b742b244e.js"></script><script defer="defer" src="../transformations/vat_rate.c7e90b2228084c918cde.js"></script><link href="../transformations/vat_rate.fd0453820f860fa87060.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="wrapper"><div class="input-group"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="input-group"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div></div><div class="row"><p class="radio-buttons">What To Do For Non-EU Country Codes</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/VAT Rate</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/vat_rate.5de9d0e807bcf50d58d7.js"></script><link href="../transformations/vat_rate.fd0453820f860fa87060.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="wrapper"><div class="input-group"><label for="input-column">Input Column</label> <select name="inputColumn" id="input-column"></select></div><div class="input-group"><label for="output-column">Output Column</label> <input name="outputColumn" id="output-column"></div></div><div class="row"><p class="radio-buttons">What To Do For Non-EU Country Codes</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave Empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Fail</label></div></div></div></div></body></html>
