+++ connect_transformations/static/vendors.b7829ba6a3fa12b6e5cb.js
@@ -1,6 +1,6 @@
 (self["webpackChunkconnect_transformations"] = self["webpackChunkconnect_transformations"] || []).push([[216],{
 
-/***/ 243:
+/***/ 164:
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
@@ -12,7 +12,7 @@ var e={705:e=>{e.exports=function e(t,n,o){function r(i,l){if(!n[i]){if(!t[i]){i
 
 /***/ }),
 
-/***/ 245:
+/***/ 824:
 /***/ ((module) => {
 
 /*!
@@ -125,11 +125,11 @@ module.exports = (function split(undef) {
 
 /***/ }),
 
-/***/ 129:
+/***/ 71:
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
 // contains, add, remove, toggle
-var indexof = __webpack_require__(906)
+var indexof = __webpack_require__(355)
 
 module.exports = ClassList
 
@@ -231,320 +231,12 @@ function isTruthy(value) {
 
 /***/ }),
 
-/***/ 906:
-/***/ ((module) => {
-
-
-var indexOf = [].indexOf;
-
-module.exports = function(arr, obj){
-  if (indexOf) return arr.indexOf(obj);
-  for (var i = 0; i < arr.length; ++i) {
-    if (arr[i] === obj) return i;
-  }
-  return -1;
-};
-
-/***/ }),
-
-/***/ 965:
-/***/ ((__unused_webpack_module, exports) => {
-
-
-function whitespace (s) {
-  return /\S/.test(s)
-}
-
-exports.START = exports.END = -1
-
-var start = exports.start = function (text, i, bound) {
-  var s = i, S = -1
-  while(s >= 0 && bound(text[s])) S = s--
-  return exports.START = S
-}
-
-var end = exports.end = function (text, i, bound) {
-  var s = i, S = -1
-  while(s < text.length && bound(text[s])) S = ++s
-  return exports.END = S
-}
-
-var word = exports.word = function (text, i, bound) {
-  bound = bound || whitespace
-  return text.substring(start(text, i, bound), end(text, i, bound))
-}
-
-exports.replace = function replace (value, text, i, bound) {
-  bound = bound || whitespace
-
-  var w = word(text, i, bound)
-  if(!w) return text
-
-  return (
-    text.substring(0, exports.START)
-  + value
-  + text.substring(exports.END)
-  )
-}
-
-
-
-/***/ }),
-
-/***/ 54:
+/***/ 789:
 /***/ ((module, __unused_webpack_exports, __webpack_require__) => {
 
-"use strict";
-
-var h = __webpack_require__(81)
-var wordBoundary = /\s/
-var bounds = __webpack_require__(965)
-
-var TextareaCaretPosition = __webpack_require__(401)
-
-var Suggester = __webpack_require__(902)
-
-module.exports = function(el, choices, options) {
-  var tcp = new TextareaCaretPosition(el)
-
-  var suggest = Suggester(choices)
-
-  options = options || {}
-
-  var stringify = options.stringify || String
-
-  var box = {
-    input: el,
-    choices: choices,
-    options: options,
-    active: false,
-    activate: activate,
-    deactivate: deactivate,
-    selection: 0,
-    filtered: [],
-
-    //get the current word
-    get: function (i) {
-      i = Number.isInteger(i) ? i : el.selectionStart - 1
-      return bounds.word(el.value, i)
-    },
-
-    //replace the current word
-    set: function (w, i) {
-      i = Number.isInteger(i) ? i : el.selectionStart - 1
-      el.value = bounds.replace(w, el.value + ' ', i)
-      el.selectionStart = el.selectionEnd = bounds.START + w.length + 1
-    },
-
-    select: function (n) {
-      this.selection = Math.max(0, Math.min(this.filtered.length, n))
-      this.update()
-    },
-    next: function () {
-      this.select(this.selection + 1)
-    },
-    prev: function () {
-      this.select(this.selection - 1)
-    },
-    suggest: function (cb) {
-      var choices, self = this
-      // extract current word
-      var word = this.get()
-      if(!word)
-        return this.deactivate(), cb()
-
-      // filter and order the list by the current word
-      this.selection = 0
-
-      var r = this.request = (this.request || 0) + 1
-      suggest(word, function (err, choices) {
-        if(err) return console.error(err)
-        if(r !== self.request) return cb()
-        if(choices) cb(null, self.filtered = choices)
-      })
-
-    },
-    reposition: function () {
-      self = this
-      if (self.filtered.length == 0)
-        return self.deactivate()
-
-      // create / update the element
-      if (self.active) {
-        self.update()
-      } else {
-        // calculate position
-        var pos = tcp.get(el.selectionStart, el.selectionEnd)
-
-        var bounds = el.getBoundingClientRect()
-        // setup
-        self.x = pos.left + bounds.left - el.scrollLeft
-        self.y = pos.top + bounds.top - el.scrollTop + 20
-        self.activate()
-      }
-    },
-    update: update,
-    complete: function (n) {
-      if(!isNaN(n)) this.select(n)
-      if (this.filtered.length) {
-        var choice = this.filtered[this.selection]
-        if (choice && choice.value) {
-          // update the text under the cursor to have the current selection's value          var v = el.value
-          this.set(stringify(choice.value))
-          // fire the suggestselect event
-          el.dispatchEvent(new CustomEvent('suggestselect', { detail: choice }))
-        }
-      }
-      this.deactivate()
-    },
-  }
-  el.addEventListener('input', oninput.bind(box))
-  el.addEventListener('keydown', onkeydown.bind(box))
-  el.addEventListener('blur', onblur.bind(box))
-  return box
-}
-
-function getItemIndex(e) {
-  for (var el = e.target; el && el != this; el = el.parentNode)
-    if (el._i != null)
-      return el._i
-}
-
-function onListMouseMove(e) {
-  this.isMouseActive = true
-}
-
-function onListMouseOver(e) {
-  // ignore mouseover triggered by list redrawn under the cursor
-  if (!this.isMouseActive) return
-
-  var i = getItemIndex(e)
-  if (i != null && i != this.selection)
-    this.select(i)
-}
-
-function onListMouseDown(e) {
-  var i = getItemIndex(e)
-  if (i != null) {
-    this.select(i)
-    this.complete()
-    // prevent blur
-    e.preventDefault()
-  }
-}
-
-function render(box) {
-  var cls = (box.options.cls) ? ('.'+box.options.cls) : ''
-  var style = { left: (box.x+'px'), position: 'fixed' }
-
-  // hang the menu above or below the cursor, wherever there is more room
-  if (box.y < window.innerHeight/2) {
-    style.top = box.y + 'px'
-  } else {
-    style.bottom = (window.innerHeight - box.y + 20) + 'px'
-  }
-
-  return h('.suggest-box'+cls, { style: style }, [
-    h('ul', {
-      onmousemove: onListMouseMove.bind(box),
-      onmouseover: onListMouseOver.bind(box),
-      onmousedown: onListMouseDown.bind(box)
-    }, renderOpts(box))
-  ])
-}
-
-function renderOpts(box) {
-  var fragment = document.createDocumentFragment()
-  for (var i=0; i < box.filtered.length; i++) {
-    var opt = box.filtered[i]
-    var tag = 'li'
-    if (i === box.selection) tag += '.selected'
-    if (opt.cls) tag += '.' + opt.cls
-    var title = null
-    var image = null
-    if(opt.showBoth){
-        title = h('strong', opt.title)
-        image = h('img', { src: opt.image })
-    } else title = opt.image ? h('img', { src: opt.image }) : h('strong', opt.title)
-    fragment.appendChild(h(tag, {_i: i}, image, ' ', [title, ' ', opt.subtitle && h('small', opt.subtitle)]))
-  }
-  return fragment
-}
-
-function activate() {
-  if (this.active)
-    return
-  this.active = true
-  this.el = render(this)
-  document.body.appendChild(this.el)
-  adjustPosition.call(this)
-}
-
-function update() {
-  if (!this.active)
-    return
-  var ul = this.el.querySelector('ul')
-  ul.innerHTML = ''
-  ul.appendChild(renderOpts(this))
-  adjustPosition.call(this)
-}
-
-function deactivate() {
-  if (!this.active)
-    return
-  this.el.parentNode.removeChild(this.el)
-  this.el = null
-  this.active = false
-}
-
-function oninput(e) {
-  var self = this
-  var word = this.suggest(function (_, suggestions) {
-    if(suggestions) self.reposition()
-  })
-}
-
-function onkeydown(e) {
-  if (this.active) {
-    // up
-    if (e.keyCode == 38) this.prev()
-    // down
-    else if (e.keyCode == 40) this.next()
-    // escape
-    else if (e.keyCode == 27) this.deactivate()
-    // enter or tab
-
-    else if (e.keyCode == 13 || e.keyCode == 9) this.complete()
-    else return //ordinary key, fall back.
-
-    e.preventDefault() //movement key, as above.
-
-    this.isMouseActive = false
-  }
-}
-
-function onblur(e) {
-  this.deactivate()
-}
-
-function adjustPosition() {
-  // move the box left to fit in the viewport, if needed
-  var width = this.el.getBoundingClientRect().width
-  var rightOverflow = this.x + width - window.innerWidth
-  var rightAdjust = Math.min(this.x, Math.max(0, rightOverflow))
-  this.el.style.left = (this.x - rightAdjust) + 'px'
-}
-
-
-
-/***/ }),
-
-/***/ 81:
-/***/ ((module, __unused_webpack_exports, __webpack_require__) => {
-
-var split = __webpack_require__(245)
-var ClassList = __webpack_require__(129)
-__webpack_require__(291)
+var split = __webpack_require__(824)
+var ClassList = __webpack_require__(71)
+__webpack_require__(525)
 
 function context () {
 
@@ -695,78 +387,23 @@ function isArray (arr) {
 
 /***/ }),
 
-/***/ 902:
+/***/ 355:
 /***/ ((module) => {
 
 
-function isObject (o) {
-  return o && 'object' === typeof o
-}
+var indexOf = [].indexOf;
 
-var isArray = Array.isArray
-
-function isFunction (f) {
-  return 'function' === typeof f
-}
-
-function compare(a, b) {
-  return compareval(a.rank, b.rank) || compareval(a.title, b.title)
-}
-
-function compareval(a, b) {
-  return a === b ? 0 : a < b ? -1 : 1
-}
-
-function suggestWord (word, choices, cb) {
-  if(isArray(choices)) {
-    //remove any non word characters and make case insensitive.
-    var wordRe = new RegExp(word.replace(/\W/g, ''), 'i')
-    cb(null, choices.map(function (opt, i) {
-      var title = wordRe.exec(opt.title)
-      var subtitle = opt.subtitle ? wordRe.exec(opt.subtitle) : null
-      var rank = (title === null ? (subtitle&&subtitle.index) : (subtitle === null ? (title&&title.index) : Math.min(title.index, subtitle.index)))
-      if (rank !== null) {
-        opt.rank = rank
-        return opt
-      }
-    }).filter(Boolean).sort(compare).slice(0, 20))
+module.exports = function(arr, obj){
+  if (indexOf) return arr.indexOf(obj);
+  for (var i = 0; i < arr.length; ++i) {
+    if (arr[i] === obj) return i;
   }
-  else if(isFunction(choices)) choices(word, cb)
-}
-
-module.exports = function (choices) {
-  if(isFunction(choices)) return choices
-
-  else if(isObject(choices) && (choices.any || isArray(choices)))
-    return function (word, cb) {
-      suggestWord(word, choices.any || choices, cb)
-    }
-  else if(isObject(choices)) {
-    var _choices = choices
-    //legacy
-    return function (word, cb) {
-      if(!choices[word[0]]) return cb()
-      suggestWord(word.substring(1), choices[word[0]], cb)
-    }
-  }
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
+  return -1;
+};
 
 /***/ }),
 
-/***/ 401:
+/***/ 969:
 /***/ ((module) => {
 
 /* jshint browser: true */
