+++ ui/src/pages/transformations/formula.js
@@ -2,7 +2,7 @@
 Copyright (c) 2023, CloudBlue LLC
 All rights reserved.
 */
-import suggestBox from 'suggest-box';
+import suggestBox from '../../libs';
 import createApp from '@cloudblueconnect/connect-ui-toolkit';
 import '@fontsource/roboto/500.css';
 import '../../../styles/index.css';
